mengamankan data
