+++ submission/dashboard/dashboard.py
@@ -1,4 +1,4 @@
-from partikulasi_polusi import load_data, page_partilukalasi_polusi_harian, page_partilukalasi_polusi_mingguan, page_partilukalasi_polusi_bulanan, page_partilukalasi_polusi_tahunan
+from submission.dashboard.partikulasi_polusi.partikulasi_polusi import load_data, page_partilukalasi_polusi_harian, page_partilukalasi_polusi_mingguan, page_partilukalasi_polusi_bulanan, page_partilukalasi_polusi_tahunan
 import streamlit as st
 
 # Load data
